feat: separate drawing from the form by route
This is how the SPA version of the app worked previously and allows it
to work without JavaScript.
